+++ CHANGELOG.md
@@ -9,13 +9,15 @@ ADDED:
 - new idea run configurations for gradle builds created
 - created file gradle.properties
 - new dependency jobj-compare in version 3.2 added
+- new dependency throw-able in version 1.2 added
+- new test method for copy run configurations files from one source project to another target project and modifies its content
 
 CHANGED:
 
 - removed idea run configurations for maven
 - moved dependency versions to file gradle.properties
 - update of silly-strings dependency version to 5.4
-- update of silly-collections dependency version to 5.8
+- update of silly-collections dependency version to 5.9
 - update of mystic-crypt dependency version to 7.3
 - update of junit dependency version to 4.13
 - update of mockito-core dependency version to 3.2.4